+++ sentry-ruby/spec/sentry/interfaces/stacktrace_spec.rb
@@ -18,14 +18,16 @@ RSpec.describe Sentry::StacktraceInterface::Frame do
     it "initializes a Frame with the correct info from the given Backtrace::Line object" do
       first_frame = Sentry::StacktraceInterface::Frame.new(configuration.project_root, lines.first)
 
-      expect(first_frame.filename).to match(/base.rb/)
+      expect(first_frame.abs_path).to eq("#{Dir.home}/.rvm/gems/activerecord/base.rb")
+      expect(first_frame.filename).to eq("#{Dir.home}/.rvm/gems/activerecord/base.rb")
       expect(first_frame.in_app).to eq(false)
       expect(first_frame.function).to eq("save")
       expect(first_frame.lineno).to eq(10)
 
       second_frame = Sentry::StacktraceInterface::Frame.new(configuration.project_root, lines.last)
 
-      expect(second_frame.filename).to match(/post.rb/)
+      expect(second_frame.abs_path).to match("#{configuration.project_root}/app/models/post.rb")
+      expect(second_frame.filename).to match("app/models/post.rb")
       expect(second_frame.in_app).to eq(true)
       expect(second_frame.function).to eq("save_user")
       expect(second_frame.lineno).to eq(5)
@@ -33,12 +35,52 @@ RSpec.describe Sentry::StacktraceInterface::Frame do
 
     it "does not strip load path when strip_backtrace_load_path is false" do
       first_frame = Sentry::StacktraceInterface::Frame.new(configuration.project_root, lines.first, false)
+      expect(first_frame.abs_path).to eq(first_frame.abs_path)
       expect(first_frame.filename).to eq(first_frame.abs_path)
       expect(first_frame.filename).to eq(raw_lines.first.split(':').first)
 
       second_frame = Sentry::StacktraceInterface::Frame.new(configuration.project_root, lines.last, false)
+      expect(second_frame.abs_path).to eq(second_frame.abs_path)
       expect(second_frame.filename).to eq(second_frame.abs_path)
       expect(second_frame.filename).to eq(raw_lines.last.split(':').first)
+    end
+
+    context "with backtrace cleaner" do
+      let(:backtrace_cleanup_callback) do
+        proc do |backtrace|
+          backtrace.map do |line|
+            if line.match(/gems/)
+              line.sub("#{Dir.home}/.rvm/gems/", "")
+            elsif line.start_with?(configuration.project_root)
+              line.sub(configuration.project_root, "")
+            else
+              line
+            end
+          end
+        end
+      end
+
+      let(:lines) do
+        Sentry::Backtrace.parse(raw_lines, configuration.project_root, configuration.app_dirs_pattern, &backtrace_cleanup_callback).lines
+      end
+
+      it "handles raw and cleaned lines properly" do
+        first_frame = Sentry::StacktraceInterface::Frame.new(configuration.project_root, lines.first)
+
+        expect(first_frame.abs_path).to eq("#{Dir.home}/.rvm/gems/activerecord/base.rb")
+        expect(first_frame.filename).to eq("activerecord/base.rb")
+        expect(first_frame.in_app).to eq(false)
+        expect(first_frame.function).to eq("save")
+        expect(first_frame.lineno).to eq(10)
+
+        second_frame = Sentry::StacktraceInterface::Frame.new(configuration.project_root, lines.last)
+
+        expect(second_frame.abs_path).to match("#{configuration.project_root}/app/models/post.rb")
+        expect(second_frame.filename).to match("app/models/post.rb")
+        expect(second_frame.in_app).to eq(true)
+        expect(second_frame.function).to eq("save_user")
+        expect(second_frame.lineno).to eq(5)
+      end
     end
   end
 end